backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
werkzeug==2.3.6
email-validator==2.0.0
